make dynamic slice/update regionsize/numsubelements 32bit
Summary:
uint16 is too small for likely sizes in some cases.
2d vertices not changed for now as they are worker vertices and their footprint
is more important.

Fixes T19900

Test Plan: CI, size benchmarks likely to require update

Reviewers: timp, samk, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: samk, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T19900

Differential Revision: https://phabricator.sourcevertex.net/D26171

+++ lib/popops/codelets/DynamicSlice1d.cpp
@@ -35,9 +35,9 @@ public:
   Input<unsigned> offset; // in \a baseT
   Input<Vector<InType, ONE_PTR>> baseT;
   Output<Vector<InType, ONE_PTR>> subT;
-  const unsigned numBaseElements;      // in the slice dimension
-  const unsigned short numSubElements; // in the slice dimension
-  const unsigned short regionSize;     // stride between slices
+  const unsigned numBaseElements; // in the slice dimension
+  const unsigned numSubElements;  // in the slice dimension
+  const unsigned regionSize;      // stride between slices
 
   IS_EXTERNAL_CODELET(!isBool<InType>());
